Move properties to its own file
