+++ src/use-cases/PopulateOffences.ts
@@ -1,26 +1,17 @@
+import {
+  ADJOURNMENT_SINE_DIE_RESULT_CODE_STRING,
+  COMMON_LAWS,
+  DONT_KNOW_VALUE,
+  ENTERED_IN_ERROR_RESULT_CODE,
+  INDICTMENT,
+  STOP_LIST,
+  TIME_RANGE
+} from "src/lib/properties"
 import type { Offence, OffenceCode } from "src/types/AnnotatedHearingOutcome"
 import type { OffenceParsedXml, ResultedCaseMessageParsedXml } from "src/types/IncomingMessage"
 import removeSeconds from "src/utils/removeSeconds"
 import { lookupResultQualifierCodeByCjsCode } from "./dataLookup"
 import PopulateOffenceResults from "./PopulateOffenceResults"
-
-const ENTERED_IN_ERROR_RESULT_CODE = 4583 // Hearing Removed
-const STOP_LIST = [
-  1000, 1505, 1509, 1510, 1511, 1513, 1514, 2069, 2501, 2505, 2507, 2508, 2509, 2511, 2514, 3501, 3502, 3503, 3504,
-  3508, 3509, 3510, 3512, 3514, 4049, 4505, 4507, 4509, 4510, 4532, 4534, 4544, 4584, 4585, 4586, 3118, 4592, 4593,
-  4594, 4595, 4596, 4597
-]
-const COMMON_LAWS = "COML"
-const INDICTMENT = "XX00"
-const DONT_KNOW_VALUE = "D"
-const ADJOURNMENT_SINE_DIE_RESULT_CODE_STRING = "2007"
-
-const ON_OR_IN = 1
-const BEFORE = 2
-const AFTER = 3
-const BETWEEN = 4
-const ON_OR_ABOUT = 5
-const ON_OR_BEFORE = 6
 
 export interface OffencesResult {
   offences: Offence[]
@@ -107,6 +98,7 @@ export default class {
     if (spiOffenceStart?.OffenceStartTime) {
       const spiOffenceStartTime = removeSeconds(spiOffenceStart.OffenceStartTime)
 
+      const { ON_OR_IN, BEFORE, AFTER, ON_OR_ABOUT, ON_OR_BEFORE, BETWEEN } = TIME_RANGE
       if ([ON_OR_IN, BEFORE, AFTER, ON_OR_ABOUT, ON_OR_BEFORE].includes(spiOffenceDateCode)) {
         offence.OffenceTime = spiOffenceStartTime
       } else if (spiOffenceDateCode === BETWEEN) {
